refactor: BookCategories를 inra 패키지로 이동
